refactor dashboard home into modular components
