+++ src/components/pages/dashboard-home/header.tsx
@@ -1,0 +1,38 @@
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select"
+import { Calendar, Filter } from "lucide-react"
+import { useDashboardHomeContext } from "@/context/dashboard-home-context"
+import { DateFilter, ShiftFilter } from "@/types/dashboard"
+
+export default function DashboardHomeHeader() {
+    const { dateFilter, setDateFilter, shiftFilter, setShiftFilter } = useDashboardHomeContext()
+
+    return (
+        <div className="flex flex-col space-y-4 md:flex-row md:items-center md:justify-between md:space-y-0">
+            <div className="flex flex-col space-y-2 sm:flex-row sm:space-y-0 sm:space-x-2">
+                <Select value={dateFilter} onValueChange={(value: DateFilter) => setDateFilter(value)}>
+                    <SelectTrigger className="w-full sm:w-48">
+                        <Calendar className="h-4 w-4 mr-2" />
+                        <SelectValue />
+                    </SelectTrigger>
+                    <SelectContent>
+                        <SelectItem value="today">Hoje</SelectItem>
+                        <SelectItem value="yesterday">Ontem</SelectItem>
+                        <SelectItem value="7days">Últimos 7 dias</SelectItem>
+                        <SelectItem value="30days">Últimos 30 dias</SelectItem>
+                    </SelectContent>
+                </Select>
+                <Select value={shiftFilter} onValueChange={(value: ShiftFilter) => setShiftFilter(value)}>
+                    <SelectTrigger className="w-full sm:w-32">
+                        <Filter className="h-4 w-4 mr-2" />
+                        <SelectValue />
+                    </SelectTrigger>
+                    <SelectContent>
+                        <SelectItem value="all">Todos</SelectItem>
+                        <SelectItem value="lunch">Almoço</SelectItem>
+                        <SelectItem value="dinner">Jantar</SelectItem>
+                    </SelectContent>
+                </Select>
+            </div>
+        </div>
+    )
+}
